Adding selectize.css to the exclude list in bower-install task.
This is because we do not want the css file bundled with selective to be included in the index.html. We are using materialize.css instead Removing it from the bower.json file of selectize is not a good option because running 'bower install' on this repo would add that file again.  This will cause a problem for another developer trying to access this repo or for a Continuous integration module to pick up the repo and run an automation script to build the resources.

Since we are not checking in the bower_components it is imperative that we find a cleaner solution.